+++ lib/coverband/collectors/trace.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Coverband
+  module Collectors
+    class Trace < Base
+    end
+  end
+end
