+++ test/unit/base_test.rb
@@ -5,32 +5,27 @@ require File.expand_path('./dog', File.dirname(__FILE__))
 
 class BaseTest < Test::Unit::TestCase
   test 'defaults to a redis store' do
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     assert_equal Coverband::Adapters::RedisStore, coverband.instance_variable_get('@store').class
   end
 
   test 'configure memory caching' do
     Coverband.configuration.memory_caching = true
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     assert_equal Coverband::Adapters::MemoryCacheStore, coverband.instance_variable_get('@store').class
     Coverband.configuration.memory_caching = false
   end
 
   test 'start should enable coverage' do
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     assert_equal false, coverband.instance_variable_get('@enabled')
     coverband.expects(:record_coverage).once
     coverband.start
     assert_equal true, coverband.instance_variable_get('@enabled')
   end
 
-  test 'extended should default to false' do
-    coverband = Coverband::Base.instance.reset_instance
-    assert_equal false, coverband.extended?
-  end
-
   test 'stop should disable coverage' do
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     assert_equal false, coverband.instance_variable_get('@enabled')
     coverband.expects(:record_coverage).once
     coverband.start
@@ -42,7 +37,7 @@ class BaseTest < Test::Unit::TestCase
 
   test 'allow for sampling with a block and enable when 100 percent sample' do
     logger = Logger.new(STDOUT)
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     coverband.instance_variable_set('@sample_percentage', 100.0)
     coverband.instance_variable_set('@verbose', true)
     coverband.instance_variable_set('@logger', logger)
@@ -55,7 +50,7 @@ class BaseTest < Test::Unit::TestCase
 
   test 'allow reporting with start stop save' do
     logger = Logger.new(STDOUT)
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     coverband.instance_variable_set('@sample_percentage', 100.0)
     coverband.instance_variable_set('@verbose', true)
     coverband.instance_variable_set('@logger', logger)
@@ -70,13 +65,13 @@ class BaseTest < Test::Unit::TestCase
 
   test 'allow reporting to redis start stop save' do
     dog_file = File.expand_path('./dog.rb', File.dirname(__FILE__))
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     coverband.instance_variable_set('@sample_percentage', 100.0)
     coverband.instance_variable_set('@verbose', true)
     Coverband.configuration.logger.stubs('info')
     store = Coverband::Adapters::RedisStore.new(Redis.new)
     coverband.instance_variable_set('@store', store)
-    store.expects(:save_report).once.with(has_entries(dog_file => { 3 => 5 }))
+    store.expects(:save_report).once.with(has_entries(dog_file => { 5 => 5 }))
     assert_equal false, coverband.instance_variable_get('@enabled')
     coverband.start
     5.times { Dog.new.bark }
@@ -86,16 +81,16 @@ class BaseTest < Test::Unit::TestCase
 
   test 'tracer should count line numbers' do
     dog_file = File.expand_path('./dog.rb', File.dirname(__FILE__))
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     coverband.start
     100.times { Dog.new.bark }
-    assert_equal 100, coverband.instance_variable_get('@file_line_usage')[dog_file][3]
+    assert_equal 100, coverband.instance_variable_get('@file_line_usage')[dog_file][5]
     coverband.stop
     coverband.save
   end
 
   test 'sample should return the result of the block' do
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     assert_equal 2, coverband.sample { 1 + 1 }
   end
 end
