+++ test/unit/configuration_test.rb
@@ -9,13 +9,13 @@ class BaseTest < Test::Unit::TestCase
 
   test 'defaults to ignore gems' do
     assert_equal Coverband.configuration.include_gems, false
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     assert_equal ['vendor', 'internal:prelude', 'schema.rb', 'gems'], coverband.instance_variable_get('@ignore_patterns')
   end
 
   test "doesn't ignore gems if include_gems = true" do
     Coverband.configuration.include_gems = true
-    coverband = Coverband::Base.instance.reset_instance
+    coverband = Coverband::Collectors::Base.instance.reset_instance
     assert_equal ['vendor', 'internal:prelude', 'schema.rb'], coverband.instance_variable_get('@ignore_patterns')
   end
 end
