+++ test/unit/reports_base_test.rb
@@ -75,7 +75,7 @@ class ReportsBaseTest < Test::Unit::TestCase
     @fake_redis.stubs(:smembers).returns(lines_hit)
     # expects to show hit counts on 1,3,6
     expected = { '/full/remote_app/path/is/a/path.rb' => [1, nil, 1, nil, nil, 1] }
-    File.stubs(:exists?).returns(true)
+    File.stubs(:exist?).returns(true)
     File.stubs(:foreach).returns(['line 1', 'line2', 'line3', 'line4', 'line5', 'line6'])
 
     assert_equal expected, Coverband::Reporters::Base.line_hash(store, key, roots)
@@ -99,7 +99,7 @@ class ReportsBaseTest < Test::Unit::TestCase
     @fake_redis.stubs(:smembers).returns(lines_hit)
     # expects to show hit counts on 1,3,6
     expected = { '/full/remote_app/path/is/a/path.rb' => [1, nil, 1, nil, nil, 1] }
-    File.stubs(:exists?).returns(true)
+    File.stubs(:exist?).returns(true)
     File.stubs(:foreach).returns(['line 1', 'line2', 'line3', 'line4', 'line5', 'line6'])
 
     assert_equal expected, Coverband::Reporters::Base.line_hash(store, key, roots)
